requirements.txt
google-generativeai
chromadb
python-dotenv
Django==5.0.3
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-cors-headers